uodated README.md renamed engine.hand_solver to engine.hand_logic added docs folder added docs/version_history.md
