+++ docs/version_history.md
@@ -1,0 +1,9 @@
+# Version History
+
+## Version 0.1.0
+**Release Date - 13/12/20**
+ - Basic project structure added
+ - github actions pipeline established for CICD
+ - poetry dependencies added to pyproject.toml
+ - build basic engine.hand_logic base class and functions sub package
+ - begun development of texas holdem implementation of engine.hand_logic class
