+++ src/pypoker/engine/hand_logic/functions/__init__.py
@@ -12,7 +12,7 @@ The public methods to access these functions will be imported here for ease of u
 with their private implementation methods
 """
 
-from pypoker.engine.hand_solver.functions.type import hand_test
-from pypoker.engine.hand_solver.functions.rank import rank_hand_type
-from pypoker.engine.hand_solver.functions.describe import describe_hand
-from pypoker.engine.hand_solver.functions.outs import find_outs_scenarios
+from pypoker.engine.hand_logic.functions.type import hand_test
+from pypoker.engine.hand_logic.functions.rank import rank_hand_type
+from pypoker.engine.hand_logic.functions.describe import describe_hand
+from pypoker.engine.hand_logic.functions.outs import find_outs_scenarios
